Change particle::property::size to double instead of integer; Added API/Doxygen documentation; added e^+ to particle datafile
git-svn-id: file:///home/olsonse/src/svn-repository/trunk/particledb@691 e877f351-6dfb-0310-a246-97bd42f90100

+++ src/particledb/Particle.h
@@ -151,8 +151,8 @@ namespace particledb { namespace Particle {
         /** Representative size of this type of particle. */
         struct size {
             typedef size super;
-            int value;
-            size(const int & sz = 1) : value(sz) {}
+            double value;
+            size(const double & sz = 1) : value(sz) {}
             std::ostream & print(std::ostream & out,
                                  const std::string & sep) const {
                 return out << "size: " << value << sep;
